r1477: Changes to the aperture module, 28feb08 (BJ with help form FS)
1 - ZERO-DIVIDE MANAGMENT
The treatment of potential 'zero-divide' now complies with standards fixed by Frank. This is transparent to the user, results are unchanged.

2 - RECTANGULAR APERTURE
The case aperture=rectangle is now correctly working, but adding internally a small rounding of the corner (too small to be visible)

3 - DISPERSION AND DELTAP
The treatment of dispersion contributions to aperture have been changed.
Now, if the user uses a sequence:
TWISS,...,  deltap=dp0
APERTURE, ... dp = dp1,
the effect of deltap is taken into account correctly, i.e. the contribution of dispersion to x and y is added a spurious dispersion proportional to deltap. This was not the case before, only the contribution to spurious dipersion proportional to dp1 was considered.
This is not changing the results for use with deltap=0 in TWISS.

On top of this change, we suppressed the beta-beating factor formerly applied to D*(dp0+dp1). THIS IS SLIGHTLY CHANGING THE RESULTS IN ALL CASES. The user shall observe slightly better n1's compared to former results.

4 - HEADER OF APERTURE TABLE AND FILE
The header now fully complies to TFS format.
The deltap of TWISS is now displayed.

5 - A NEW VARIABLE IN THE BEAM STRUCTURE
To ease matching with aperture, a variable n1min is added to the BEAM data structure. It can be retrieved as BEAM%myseq->n1min if the sequence 'myseq' was in use when an APERTURE command was executed. It is initialised to n1min=-1.0.

6 - TERMINAL OUTPUT MADE CLEANER
undesirable terminal output line (enter survey module, etc) are supressed