r2544: In routine the character string tmp:
char tmp[26];

has been extended to 26 from 16. This was needed since in the sprintf command:

      sprintf(tmp, v_format("%F"), e_deltap);

it had to be long enough to hold a double precision number. This is only an 
issue on the later versions of gcc on the MAC. Under Linux none of the gcc 
versions did complain about anything. This might be a pseudo bug but needed
to work on MAC.