Fix growth of zero-length arrays
In many places the `grow_XXX_array` functions are used to allocate an
additional array element directly before writing into the supposedly
allocated array index without further range-checks.

Now, if an array was initialized to zero length it was never grown,
meaning that we would write after the end of the allocated buffer.

This just caught me with the `select->at` parameter which has to have
zero length by default (as defined in mad_dict.c). If you specify one or
two values for `at`, everything *seemed* to work as expected. However,
`decode_par` did not properly grow the double array and was writing the
values after the end of the buffer. When specifying a third argument,
the problem manifested by writing a NUL into the first byte of the
`flag` buffer and hence not anymore executing any selection routine.

I need at least grow_double_array/grow_expr_list to make this work.

Note: we could alternatively fix this by always allocating at least an
`max=1` array to begin with.