Manage drift numbers per sequence and fix memory leak
This fixes the issue of increasing the drift numbers each time a
sequence is USEd due to using a global drift counter and the memory leak
due to adding new implicit drifts to `element_list` and `drift_list` on
every USE but never freeing them when the sequence is reUSEd. The bad
behaviour was introduced by disabling the sharing of drifts.

NOTE: this has the serious consequence that implicit drifts can not be
accessed with any command that uses `element_list`.