Handle bare parameters in `command_par_string_user`
This will allow further simplification in a fair number of cases where
`call_def` is accessed explicitly throughout the code base.

I don't think there are any places where you would want to receive a
NULL if the command defines a "bareword value" and the user passes a
parameter as bareword. If it turns out that some logic depended on that,
it is probably broken.